move all components in "components" folder
