+++ src/app/@core/modules/alert.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { AlertsnackbarComponent } from '../../alertsnackbar/alertsnackbar.component';
+import { AlertsnackbarComponent } from '../../components/alertsnackbar/alertsnackbar.component';
 
 @NgModule({
   imports: [CommonModule],
