+++ src/app/components/edit-substance-data/edit-substance-data.component.ts
@@ -11,11 +11,11 @@ import {
   SubstanceData,
   Unit,
   unitGroups,
-} from '../@core/models/substances.model';
-import { compareArrays } from '../@core/utils/compare';
-import { GlobalModel } from '../@core/models/global.model';
-import { LocalizedStrings } from '../@core/services/i18n/i18n.interface';
-import Logger from '../@core/utils/logger';
+} from '../../@core/models/substances.model';
+import { compareArrays } from '../../@core/utils/compare';
+import { GlobalModel } from '../../@core/models/global.model';
+import { LocalizedStrings } from '../../@core/services/i18n/i18n.interface';
+import Logger from '../../@core/utils/logger';
 import { YesNoDialogComponent } from '../yes-no-dialog/yes-no-dialog.component';
 
 const logger = new Logger('edit-substance-data');
