+++ src/app/components/loading/loading.component.scss
