+++ src/app/components/loading/loading.component.ts
