+++ src/app/components/manual/manual.component.html
