+++ src/app/components/manual/manual.component.scss
