+++ src/app/components/manual/manual.component.spec.ts
