+++ src/app/components/menubar/menubar.component.ts
@@ -3,20 +3,20 @@ import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 import { first, map, switchMap } from 'rxjs/operators';
 import { MatDialog } from '@angular/material/dialog';
 
-import { AlertService } from '../@core/services/alertsnackbar/altersnackbar.service';
-import { CaBr2Document } from '../@core/services/loadSave/loadSave.model';
-import { compareArrays } from '../@core/utils/compare';
-import { DialogFilter } from '../@core/services/native/tauri.service';
-import { GlobalModel } from '../@core/models/global.model';
-import { IConfigService } from '../@core/services/config/config.interface';
-import { ILoadSaveService } from '../@core/services/loadSave/loadSave.interface';
-import { INativeService } from '../@core/services/native/native.interface';
-import { LocalizedStrings } from '../@core/services/i18n/i18n.interface';
-import Logger from '../@core/utils/logger';
+import { AlertService } from '../../@core/services/alertsnackbar/altersnackbar.service';
+import { CaBr2Document } from '../../@core/services/loadSave/loadSave.model';
+import { compareArrays } from '../../@core/utils/compare';
+import { DialogFilter } from '../../@core/services/native/tauri.service';
+import { GlobalModel } from '../../@core/models/global.model';
+import { IConfigService } from '../../@core/services/config/config.interface';
+import { ILoadSaveService } from '../../@core/services/loadSave/loadSave.interface';
+import { INativeService } from '../../@core/services/native/native.interface';
+import { LocalizedStrings } from '../../@core/services/i18n/i18n.interface';
+import Logger from '../../@core/utils/logger';
 import { ManualComponent } from '../manual/manual.component';
 import { ReportBugComponent } from '../report-bug/report-bug.component';
 import { SettingsComponent } from '../settings/settings.component';
-import TEMPLATES from '../../assets/docsTemplate.json';
+import TEMPLATES from '../../../assets/docsTemplate.json';
 import { YesNoDialogComponent } from '../yes-no-dialog/yes-no-dialog.component';
 
 const logger = new Logger('menubar');
