+++ src/app/components/modifiable-string-list/modifiable-string-list.component.scss
