+++ src/app/components/preview/preview.component.scss
