+++ src/app/components/report-bug/report-bug.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { MatDialogRef } from '@angular/material/dialog';
 
-import { GlobalModel } from '../@core/models/global.model';
-import { INativeService } from '../@core/services/native/native.interface';
-import { LocalizedStrings } from '../@core/services/i18n/i18n.interface';
+import { GlobalModel } from '../../@core/models/global.model';
+import { INativeService } from '../../@core/services/native/native.interface';
+import { LocalizedStrings } from '../../@core/services/i18n/i18n.interface';
 
 @Component({
   selector: 'app-report-bug',
