+++ src/app/components/search/search-dialog/search-dialog.component.scss
