+++ src/app/components/search/search.component.scss
