+++ src/app/components/search/selected-search/selected-search.component.html
