+++ src/app/components/search/selected-search/selected-search.component.spec.ts
