+++ src/app/components/yes-no-dialog/yes-no-dialog.component.scss
